Improved code formatting (braces for control structures)
git-svn-id: http://dompdf.googlecode.com/svn/trunk/dompdf@499 8e70de24-3a84-11de-b438-597f59cd7555

+++ include/cpdf_adapter.cls.php
@@ -159,12 +159,15 @@ class CPDF_Adapter implements Canvas {
    */
   function __construct($paper = "letter", $orientation = "portrait") {
 
-    if ( is_array($paper) )
+    if ( is_array($paper) ) {
       $size = $paper;
-    else if ( isset(self::$PAPER_SIZES[mb_strtolower($paper)]) )
+    }
+    else if ( isset(self::$PAPER_SIZES[mb_strtolower($paper)]) ) {
       $size = self::$PAPER_SIZES[mb_strtolower($paper)];
-    else
+    }
+    else {
       $size = self::$PAPER_SIZES["letter"];
+    }
 
     if ( mb_strtolower($orientation) === "landscape" ) {
       list($size[2], $size[3]) = array($size[3], $size[2]);
@@ -196,8 +199,7 @@ class CPDF_Adapter implements Canvas {
     foreach ($this->_image_cache as $img) {
       //debugpng
       if (DEBUGPNG) print '[__destruct unlink '.$img.']';
-      if (!DEBUGKEEPTEMP)
-        unlink($img);
+      if (!DEBUGKEEPTEMP) unlink($img);
     }
   }
   
@@ -559,8 +561,9 @@ class CPDF_Adapter implements Canvas {
     $this->_set_stroke_color($color);
     
     // Adjust y values
-    for ( $i = 1; $i < count($points); $i += 2)
+    for ( $i = 1; $i < count($points); $i += 2) {
       $points[$i] = $this->y($points[$i]);
+    }
     
     $this->_pdf->polygon($points, count($points) / 2, $fill);
   }
@@ -571,8 +574,9 @@ class CPDF_Adapter implements Canvas {
     $this->_set_fill_color($color);
     $this->_set_stroke_color($color);
 
-    if ( !$fill && isset($width) )
+    if ( !$fill && isset($width) ) {
       $this->_set_line_style($width, "round", "round", $style);
+    }
 
     $this->_pdf->ellipse($x, $this->y($y), $r1, 0, 0, 8, 0, 360, 1, $fill);
   }
@@ -587,18 +591,18 @@ class CPDF_Adapter implements Canvas {
 
     switch ($type) {
     case IMAGETYPE_JPEG:
-      if (DEBUGPNG)  print '!!!jpg!!!';
+      if (DEBUGPNG) print '!!!jpg!!!';
       $this->_pdf->addJpegFromFile($img, $x, $this->y($y) - $h, $w, $h);
       break;
       
     case IMAGETYPE_GIF:
     case IMAGETYPE_BMP:
-      if (DEBUGPNG)  print '!!!bmp or gif!!!';
+      if (DEBUGPNG) print '!!!bmp or gif!!!';
       // @todo use cache for BMP and GIF
       $img = $this->_convert_gif_bmp_to_png($img, $type);
 
     case IMAGETYPE_PNG:
-      if (DEBUGPNG)  print '!!!png!!!';
+      if (DEBUGPNG) print '!!!png!!!';
 
       $this->_pdf->addPngFromFile($img, $x, $this->y($y) - $h, $w, $h);
       break;
@@ -685,10 +689,12 @@ class CPDF_Adapter implements Canvas {
     if ( strpos($url, '#') === 0 ) {
       // Local link
       $name = substr($url,1);
-      if ( $name )
+      if ( $name ) {
         $this->_pdf->addInternalLink($name, $x, $y, $x + $width, $y + $height);
-
-    } else {
+      }
+      
+    }
+    else {
       $this->_pdf->addLink(rawurldecode($url), $x, $y, $x + $width, $y + $height);
     }
   }
@@ -784,8 +790,9 @@ class CPDF_Adapter implements Canvas {
    */
   protected function _add_page_text() {
     
-    if ( !count($this->_page_text) )
+    if ( !count($this->_page_text) ) {
       return;
+    }
 
     $page_number = 1;
     $eval = null;
@@ -797,19 +804,18 @@ class CPDF_Adapter implements Canvas {
         extract($pt);
 
         switch ($_t) {
+          case "text":
+            $text = str_replace(array("{PAGE_NUM}","{PAGE_COUNT}"),
+                                array($page_number, $this->_page_count), $text);
+            $this->text($x, $y, $text, $font, $size, $color, $word_space, $char_space, $angle);
+            break;
             
-        case "text":
-        $text = str_replace(array("{PAGE_NUM}","{PAGE_COUNT}"),
-                            array($page_number, $this->_page_count), $text);
-        $this->text($x, $y, $text, $font, $size, $color, $word_space, $char_space, $angle);
-          break;
-          
-        case "script":
-          if (!$eval) {
-            $eval = new PHP_Evaluator($this);
-          }
-          $eval->evaluate($code, array('PAGE_NUM' => $page_number, 'PAGE_COUNT' => $this->_page_count));
-          break;
+          case "script":
+            if ( !$eval ) {
+              $eval = new PHP_Evaluator($this);
+            }
+            $eval->evaluate($code, array('PAGE_NUM' => $page_number, 'PAGE_COUNT' => $this->_page_count));
+            break;
         }
       }
